REORGANIZE FILE TREE
- delete Hero.test.tsx (practice/test file)

- create dirs for components in app/components
- move hero.tsx and navbar.tsx to their respective dirs

- in hero.tsx:
- minor wording change in `Image` component's `alt` attr

- in page.tsx:
- update imports for Navbar and Hero
